[PLUGIN-170] Move BigtableOutputFormat to sink directory
